+++ changelog.md
@@ -37,6 +37,9 @@ Codion Change Log
 - ConditionPanel.conditionView() and TableConditionPanel.conditionView() renamed view().
 - FilterTableCellRenderer.CellColors replaced with ColorProvider, Builder.foreground() and background() added, related refactoring.
 - FilterTableCellRenderer.Builder, column identifier type parameter removed.
+- FilterTableCellRenderer.Factory.create() parameter now column identifier instead of column.
+- FilterTableCellRenderer.Builder, column value type parameter added.
+- FilterTableCellEditor.Factory.create() parameter now column identifier instead of column.
 ### is.codion.framework.domain
 - Entity.copy(), copyBuilder() and deepCopy() moved to Entity.Copy, renamed mutable(), builder() and deep() respectively, immutable() added.
 - ImmutableEntity bug fixed, StackOverflowError when creating an immutable entity with cyclical foreign key dependencies, some refactoring.